@@ -1,10 +1,10 @@
 name: color_picker_example
 description: Two examples of how to use the FlexColorPicker package.
-version: 3.3.1
+version: 3.4.0
 publish_to: 'none'
 environment:
   sdk: '>=2.19.0 <4.0.0'
-  flutter: '>=3.7.0'
+  flutter: '>=3.19.0'
 
 dependencies:
   # https://pub.dev/packages/flex_color_picker
@@ -15,7 +15,7 @@ dependencies:
     sdk: flutter
 
   # https://pub.dev/packages/flutter_riverpod
-  flutter_riverpod: ^2.4.9
+  flutter_riverpod: ^2.4.10
 
   # https://pub.dev/packages/google_fonts
   google_fonts: ^6.1.0
@@ -28,7 +28,7 @@ dependencies:
 
   # https://pub.dev/packages/url_launcher
   # Used for launching a WEB URL (by Google flutter.dev).
-  url_launcher: ^6.2.3
+  url_launcher: ^6.2.5
 
 dev_dependencies:
   flutter_test:
