@@ -13,10 +13,9 @@ import 'functions/picker_functions.dart';
 import 'models/color_picker_action_buttons.dart';
 import 'models/color_picker_copy_paste_behavior.dart';
 import 'models/color_picker_type.dart';
-import 'universal_widgets/if_wrapper.dart';
 import 'widgets/color_code_field.dart';
 import 'widgets/color_picker_toolbar.dart';
-import 'widgets/context_copy_paste_menu.dart';
+import 'widgets/copy_paste_handler.dart';
 import 'widgets/main_colors.dart';
 import 'widgets/opacity/opacity_slider.dart';
 import 'widgets/picker_selector.dart';
@@ -81,6 +80,7 @@ class ColorPicker extends StatefulWidget {
       ColorPickerType.accent: true,
       ColorPickerType.bw: false,
       ColorPickerType.custom: false,
+      ColorPickerType.customSecondary: false,
       ColorPickerType.wheel: false,
     },
     this.enableShadesSelection = true,
@@ -90,6 +90,8 @@ class ColorPicker extends StatefulWidget {
     this.crossAxisAlignment = CrossAxisAlignment.center,
     this.padding = const EdgeInsets.all(16),
     this.columnSpacing = 8,
+    this.toolbarSpacing,
+    this.shadesSpacing,
     // Opacity slider
     this.enableOpacity = false,
     this.opacityTrackHeight = 36,
@@ -154,13 +156,24 @@ class ColorPicker extends StatefulWidget {
       ColorPickerType.bw: _selectBlackWhiteLabel,
       ColorPickerType.both: _selectBothLabel,
       ColorPickerType.custom: _selectCustomLabel,
+      ColorPickerType.customSecondary: _selectCustomSecondaryLabel,
       ColorPickerType.wheel: _selectWheelAnyLabel,
     },
-    // Custom color swatches and name map for the custom color swatches.
+    // Custom color, swatches and name map for the custom color swatches.
     this.customColorSwatchesAndNames = const <ColorSwatch<Object>, String>{},
+    this.customSecondaryColorSwatchesAndNames =
+        const <ColorSwatch<Object>, String>{},
     //
   })  : assert(columnSpacing >= 0 && columnSpacing <= 300,
             'The picker item column spacing must be from 0 to max 300 dp.'),
+        assert(
+            toolbarSpacing == null ||
+                (toolbarSpacing >= 0 && toolbarSpacing <= 300),
+            'The spacing must be null or from 0 to max 300 dp.'),
+        assert(
+            shadesSpacing == null ||
+                (shadesSpacing >= 0 && shadesSpacing <= 300),
+            'The spacing must be null or from 0 to max 300 dp.'),
         assert(spacing >= 0 && spacing <= 50,
             'The picker item spacing must be from 0 to max 50 dp.'),
         assert(runSpacing >= 0 && runSpacing <= 50,
@@ -284,6 +297,18 @@ class ColorPicker extends StatefulWidget {
   /// Defaults to 8 dp. Must be from 0 to 300 dp.
   final double columnSpacing;
 
+  /// Vertical spacing below the top toolbar header and action buttons.
+  ///
+  /// If not defined, defaults to [columnSpacing].
+  /// Must be null or from 0 to 300 dp.
+  final double? toolbarSpacing;
+
+  /// Vertical spacing below the Material-2 based color shades palette.
+  ///
+  /// If not defined, defaults to [columnSpacing].
+  /// Must be null or from 0 to 300 dp.
+  final double? shadesSpacing;
+
   /// Enable the opacity control for the color value.
   ///
   /// Set to true to allow users to control the opacity value of the
@@ -293,6 +318,11 @@ class ColorPicker extends StatefulWidget {
   /// When enabled, the opacity value is not returned as a separate value,
   /// it is returned in the alpha channel of the returned ARGB color value, in
   /// the onColor callbacks.
+  ///
+  /// When false, colors that has any other alpha value than 0xFF are changed
+  /// to 0xFF. To allow custom colors and pasted in color values without
+  /// setting [enableOpacity] to true and showing the opacity slider, set
+  /// [enableTransparentCustomColors] to true.
   ///
   /// Defaults to false.
   final bool enableOpacity;
@@ -346,7 +376,7 @@ class ColorPicker extends StatefulWidget {
   /// for main and swatch shades indicator items.
   ///
   /// If false, the tonal color items will be smaller and auto sized for the
-  /// palette to be same width as the Material Color palette.
+  /// palette to be same width as the Material-2 Color palette.
   ///
   /// Defaults to false. The color boxes are smaller, but length of their
   /// items is the same as MaterialColor swatch. You may prefer true to get
@@ -662,6 +692,7 @@ class ColorPicker extends StatefulWidget {
   ///  * [ColorPickerType.accent] : 'Primary'
   ///  * [ColorPickerType.bw] : 'Black & White'
   ///  * [ColorPickerType.custom] : 'Custom'
+  ///  * [ColorPickerType.customSecondary] : 'Option'
   ///  * [ColorPickerType.wheel] : 'Wheel'
   final Map<ColorPickerType, String> pickerTypeLabels;
 
@@ -675,6 +706,17 @@ class ColorPicker extends StatefulWidget {
   /// Defaults to an empty map. If the map is empty, the custom colors picker
   /// will not be shown even if it is enabled in [pickersEnabled].
   final Map<ColorSwatch<Object>, String> customColorSwatchesAndNames;
+
+  /// Color swatch to name map, with custom swatches and their names.
+  ///
+  /// Used to provide custom [ColorSwatch] objects to the custom color picker,
+  /// including their custom name label. These colors, their swatch shades
+  /// and names, are shown and used when the picker type
+  /// [ColorPickerType.customSecondary] option is enabled in the color picker.
+  ///
+  /// Defaults to an empty map. If the map is empty, the custom colors picker
+  /// will not be shown even if it is enabled in [pickersEnabled].
+  final Map<ColorSwatch<Object>, String> customSecondaryColorSwatchesAndNames;
 
   /// English default label for picker with both primary and accent colors.
   static const String _selectBothLabel = 'Primary & Accent';
@@ -690,6 +732,9 @@ class ColorPicker extends StatefulWidget {
 
   /// English default label for picker with custom defined colors.
   static const String _selectCustomLabel = 'Custom';
+
+  /// English default label for picker with custom secondary defined colors.
+  static const String _selectCustomSecondaryLabel = 'Option';
 
   /// English default label for the HSV wheel picker that can select any color.
   static const String _selectWheelAnyLabel = 'Wheel';
@@ -906,93 +951,207 @@ class ColorPicker extends StatefulWidget {
     // Make the dialog OK button.
     final String okButtonLabel =
         actionButtons.dialogOkButtonLabel ?? translate.okButtonLabel;
-    final Widget okButtonContent = Wrap(
-      alignment: WrapAlignment.center,
-      crossAxisAlignment: WrapCrossAlignment.center,
-      children: <Widget>[
-        if (actionButtons.dialogActionIcons)
-          Padding(
-            padding: const EdgeInsetsDirectional.only(end: 4),
-            child: Icon(actionButtons.okIcon),
-          ),
-        Text(okButtonLabel),
-      ],
-    );
+    final Text okButtonContent = Text(okButtonLabel);
     Widget okButton;
     switch (actionButtons.dialogOkButtonType) {
       case ColorPickerActionButtonType.text:
-        okButton = TextButton(
-          onPressed: () {
-            Navigator.of(context, rootNavigator: actionButtons.useRootNavigator)
-                .pop(true);
-          },
-          child: okButtonContent,
-        );
-        break;
+        okButton = actionButtons.dialogActionIcons
+            ? TextButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                icon: Icon(actionButtons.okIcon),
+                label: okButtonContent,
+              )
+            : TextButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                child: okButtonContent,
+              );
       case ColorPickerActionButtonType.outlined:
-        okButton = OutlinedButton(
-          onPressed: () {
-            Navigator.of(context, rootNavigator: actionButtons.useRootNavigator)
-                .pop(true);
-          },
-          child: okButtonContent,
-        );
-        break;
+        okButton = actionButtons.dialogActionIcons
+            ? OutlinedButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                icon: Icon(actionButtons.okIcon),
+                label: okButtonContent,
+              )
+            : OutlinedButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                child: okButtonContent,
+              );
       case ColorPickerActionButtonType.elevated:
-        okButton = ElevatedButton(
-          onPressed: () {
-            Navigator.of(context, rootNavigator: actionButtons.useRootNavigator)
-                .pop(true);
-          },
-          child: okButtonContent,
-        );
-        break;
+        okButton = actionButtons.dialogActionIcons
+            ? ElevatedButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                icon: Icon(actionButtons.okIcon),
+                label: okButtonContent,
+              )
+            : ElevatedButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                child: okButtonContent,
+              );
+      case ColorPickerActionButtonType.filled:
+        okButton = actionButtons.dialogActionIcons
+            ? FilledButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                icon: Icon(actionButtons.okIcon),
+                label: okButtonContent,
+              )
+            : FilledButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                child: okButtonContent,
+              );
+      case ColorPickerActionButtonType.filledTonal:
+        okButton = actionButtons.dialogActionIcons
+            ? FilledButton.tonalIcon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                icon: Icon(actionButtons.okIcon),
+                label: okButtonContent,
+              )
+            : FilledButton.tonal(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(true);
+                },
+                child: okButtonContent,
+              );
     }
 
     // Make the dialog OK button.
     final String cancelButtonLabel =
         actionButtons.dialogCancelButtonLabel ?? translate.cancelButtonLabel;
-    final Widget cancelButtonContent = Wrap(
-      alignment: WrapAlignment.center,
-      crossAxisAlignment: WrapCrossAlignment.center,
-      children: <Widget>[
-        if (actionButtons.dialogActionIcons)
-          Padding(
-            padding: const EdgeInsetsDirectional.only(end: 4),
-            child: Icon(actionButtons.closeIcon),
-          ),
-        Text(cancelButtonLabel),
-      ],
-    );
+    final Widget cancelButtonContent = Text(cancelButtonLabel);
     Widget cancelButton;
     switch (actionButtons.dialogCancelButtonType) {
       case ColorPickerActionButtonType.text:
-        cancelButton = TextButton(
-          onPressed: () {
-            Navigator.of(context, rootNavigator: actionButtons.useRootNavigator)
-                .pop(false);
-          },
-          child: cancelButtonContent,
-        );
-        break;
+        cancelButton = actionButtons.dialogActionIcons
+            ? TextButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                icon: Icon(actionButtons.closeIcon),
+                label: cancelButtonContent,
+              )
+            : TextButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                child: cancelButtonContent,
+              );
       case ColorPickerActionButtonType.outlined:
-        cancelButton = OutlinedButton(
-          onPressed: () {
-            Navigator.of(context, rootNavigator: actionButtons.useRootNavigator)
-                .pop(false);
-          },
-          child: cancelButtonContent,
-        );
-        break;
+        cancelButton = actionButtons.dialogActionIcons
+            ? OutlinedButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                icon: Icon(actionButtons.closeIcon),
+                label: cancelButtonContent,
+              )
+            : OutlinedButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                child: cancelButtonContent,
+              );
       case ColorPickerActionButtonType.elevated:
-        cancelButton = ElevatedButton(
-          onPressed: () {
-            Navigator.of(context, rootNavigator: actionButtons.useRootNavigator)
-                .pop(false);
-          },
-          child: cancelButtonContent,
-        );
-        break;
+        cancelButton = actionButtons.dialogActionIcons
+            ? ElevatedButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                icon: Icon(actionButtons.closeIcon),
+                label: cancelButtonContent,
+              )
+            : ElevatedButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                child: cancelButtonContent,
+              );
+      case ColorPickerActionButtonType.filled:
+        cancelButton = actionButtons.dialogActionIcons
+            ? FilledButton.icon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                icon: Icon(actionButtons.closeIcon),
+                label: cancelButtonContent,
+              )
+            : FilledButton(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                child: cancelButtonContent,
+              );
+      case ColorPickerActionButtonType.filledTonal:
+        cancelButton = actionButtons.dialogActionIcons
+            ? FilledButton.tonalIcon(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                icon: Icon(actionButtons.closeIcon),
+                label: cancelButtonContent,
+              )
+            : FilledButton.tonal(
+                onPressed: () {
+                  Navigator.of(context,
+                          rootNavigator: actionButtons.useRootNavigator)
+                      .pop(false);
+                },
+                child: cancelButtonContent,
+              );
     }
 
     // False if dialog cancelled, true if color selected
@@ -1023,9 +1182,9 @@ class ColorPicker extends StatefulWidget {
           ? <Widget>[
               if (okIsLeft) ...<Widget>[
                 okButton,
-                cancelButton,
+                if (!actionButtons.dialogActionOnlyOkButton) cancelButton,
               ] else ...<Widget>[
-                cancelButton,
+                if (!actionButtons.dialogActionOnlyOkButton) cancelButton,
                 okButton,
               ]
             ]
@@ -1071,7 +1230,7 @@ class ColorPicker extends StatefulWidget {
         from: context,
         to: Navigator.of(
           context,
-          rootNavigator: true,
+          rootNavigator: actionButtons.useRootNavigator,
         ).context,
       );
       if (useSafeArea) {
@@ -1123,10 +1282,15 @@ class _ColorPickerState extends State<ColorPicker> {
   // Which picker are we using now.
   late ColorPickerType _activePicker;
 
-  // Current selected color
+  // Current selected color, it never has opacity. The slider sets opacity to
+  // in the _opacity property.
   late Color _selectedColor;
 
-  // Current color opacity value
+  // A color that is tapped on in indicators. This color may have opacity if
+  // the source color has opacity.
+  late Color _tappedColor;
+
+  // Current color opacity value, this is the opacity set by the slider.
   late double _opacity;
 
   // Color picker indicator selected item should request focus.
@@ -1158,6 +1322,9 @@ class _ColorPickerState extends State<ColorPicker> {
   // sub palette, it is is not selected.
   bool _tonalOperated = false;
 
+  // Set to true when widget update triggered via internal state change.
+  bool _fromInternal = false;
+
   // Color wheel picker should request focus.
   bool _wheelShouldFocus = false;
 
@@ -1188,11 +1355,10 @@ class _ColorPickerState extends State<ColorPicker> {
   @override
   void initState() {
     super.initState();
-    // Set selected color to the widget constructor provided start color,
-    // if opacity is not enabled, override alpha to fully opaque.
-    _selectedColor =
-        widget.enableOpacity ? widget.color : widget.color.withAlpha(0xFF);
-    // Get current opacity of passed in color, if opacity is enabled.
+    // We always remove alpha from incoming color.
+    _selectedColor = widget.color.withAlpha(0xFF);
+    _tappedColor = widget.color;
+    // Opacity is captured in _opacity if enabled.
     _opacity = widget.enableOpacity ? widget.color.opacity : 1;
     // Picker labels, use english fallbacks if none provided.
     _pickerLabels = <ColorPickerType, String>{
@@ -1207,6 +1373,9 @@ class _ColorPickerState extends State<ColorPicker> {
           ColorPicker._selectBlackWhiteLabel,
       ColorPickerType.custom: widget.pickerTypeLabels[ColorPickerType.custom] ??
           ColorPicker._selectCustomLabel,
+      ColorPickerType.customSecondary:
+          widget.pickerTypeLabels[ColorPickerType.customSecondary] ??
+              ColorPicker._selectCustomSecondaryLabel,
       ColorPickerType.wheel: widget.pickerTypeLabels[ColorPickerType.wheel] ??
           ColorPicker._selectWheelAnyLabel,
     };
@@ -1217,9 +1386,11 @@ class _ColorPickerState extends State<ColorPicker> {
       ColorPickerType.accent: ColorTools.accentColors,
       ColorPickerType.bw: ColorTools.blackAndWhite,
       ColorPickerType.custom: widget.customColorSwatchesAndNames.keys.toList(),
+      ColorPickerType.customSecondary:
+          widget.customSecondaryColorSwatchesAndNames.keys.toList(),
       ColorPickerType.wheel: <ColorSwatch<Object>>[
         // Make a swatch of the selected color in the wheel.
-        ColorTools.primarySwatch(_selectedColor.withAlpha(0xFF))
+        ColorTools.primarySwatch(_selectedColor)
       ],
     };
     // Enabled color pickers, with defaults if not specified.
@@ -1235,6 +1406,10 @@ class _ColorPickerState extends State<ColorPicker> {
           // Custom picker is always disabled if no custom swatches are given.
           (widget.pickersEnabled[ColorPickerType.custom] ?? false) &&
               widget.customColorSwatchesAndNames.isNotEmpty,
+      ColorPickerType.customSecondary:
+          // Custom secondary is always disabled if no custom swatches are given
+          (widget.pickersEnabled[ColorPickerType.customSecondary] ?? false) &&
+              widget.customSecondaryColorSwatchesAndNames.isNotEmpty,
       ColorPickerType.wheel:
           widget.pickersEnabled[ColorPickerType.wheel] ?? false,
     };
@@ -1248,6 +1423,7 @@ class _ColorPickerState extends State<ColorPicker> {
     // Always update tonal when ColorPicker is initialized.
     _tonalShouldUpdate = true;
     _tonalOperated = false;
+    _fromInternal = false;
     // If there are no shade or tonal colors displayed, the wheel must
     // focus on init.
     _wheelShouldFocus =
@@ -1263,23 +1439,46 @@ class _ColorPickerState extends State<ColorPicker> {
 
   @override
   void didUpdateWidget(ColorPicker oldWidget) {
+    super.didUpdateWidget(oldWidget);
     if (_debug) {
-      debugPrint('didUpdateWidget called **********************************');
+      debugPrint('didUpdateWidget called: fromInternal: $_fromInternal'
+          ' ******************************');
     }
     // Set to true if a change was done where we need to find the picker again.
     bool shouldFindPickerAndSwatch = false;
-
     // Opacity enable/disable changed, update selected color and opacity.
     if (widget.enableOpacity != oldWidget.enableOpacity) {
-      if (_debug) {
-        debugPrint('didUpdateWidget enableOpacity = '
-            '${widget.enableOpacity == oldWidget.enableOpacity}');
-      }
-      _selectedColor =
-          widget.enableOpacity ? widget.color : widget.color.withAlpha(0xFF);
       _opacity = widget.enableOpacity ? widget.color.opacity : 1;
+      if (_debug) {
+        debugPrint('didUpdateWidget changed: enableOpacity = '
+            '${widget.enableOpacity == oldWidget.enableOpacity}'
+            ' opacity=$_opacity');
+      }
     }
-
+    // The color was updated externally, update to new color and find picker.
+    if (widget.color != _selectedColor || widget.color != _tappedColor) {
+      if (_debug) {
+        debugPrint('didUpdateWidget changed color: '
+            'color=${widget.color} selectedColor=$_selectedColor');
+      }
+      _selectedColor = widget.color.withAlpha(0xFF);
+      _opacity = widget.enableOpacity ? widget.color.opacity : 1;
+      // Make a swatch too be to find it on wheel, if color is there.
+      _typeToSwatchMap[ColorPickerType.wheel] = <ColorSwatch<Object>>[
+        ColorTools.createPrimarySwatch(_selectedColor),
+      ];
+      if (!_fromInternal) {
+        _tappedColor = _selectedColor;
+        // Wheel and edit needs to update.
+        _wheelShouldUpdate = true;
+        _editShouldUpdate = true;
+        // Tonal picker should update from external change.
+        _tonalShouldUpdate = true;
+        _tonalOperated = false;
+        // We need to find the right picker again.
+        shouldFindPickerAndSwatch = true;
+      }
+    }
     // Picker labels map changed, update used one, with its default fallbacks.
     if (!mapEquals(widget.pickerTypeLabels, oldWidget.pickerTypeLabels)) {
       if (_debug) {
@@ -1302,32 +1501,32 @@ class _ColorPickerState extends State<ColorPicker> {
         ColorPickerType.custom:
             widget.pickerTypeLabels[ColorPickerType.custom] ??
                 ColorPicker._selectCustomLabel,
+        ColorPickerType.customSecondary:
+            widget.pickerTypeLabels[ColorPickerType.customSecondary] ??
+                ColorPicker._selectCustomSecondaryLabel,
         ColorPickerType.wheel: widget.pickerTypeLabels[ColorPickerType.wheel] ??
             ColorPicker._selectWheelAnyLabel,
       };
     }
-
     // Pickers customColorSwatchesAndNames map changed, or pickersEnabled map
     // changed, they depend on each other, so we always update state of both.
     if (widget.customColorSwatchesAndNames.toString() !=
             oldWidget.customColorSwatchesAndNames.toString() ||
         !mapEquals(widget.pickersEnabled, oldWidget.pickersEnabled)) {
-      // TODO(rydmike): Investigate the mapEquals issue.
       // In above un-equality check, the mapEquals, or with map != operator,
       // does not work if you provide a map made with createPrimarySwatch or
-      // createAccentSwatch. It should, not sure why it does not. The Wheel
-      // Picker does not work entirely as intended if the above if is
-      // evaluated incorrectly. The issue will be seen via that it will not
-      // keep the Wheel picker active when selecting known swatch color in it,
-      // it will instead move to the picker where it exists. This is not
-      // desired, we want to stay on the Wheel in this case. If you see the
-      // wrong behavior it is due to the  above IF being evaluated incorrectly.
+      // createAccentSwatch. The Wheel Picker does not work entirely as
+      // intended if the above IF is evaluated incorrectly. The wrong result
+      // will be noticed seen as that Wheel picker will not be kept active
+      // when selecting known swatch color in it. It will instead move to
+      // the picker to Swatch where the color exists. This is not desired.
+      // We want to stay on the Wheel in this case. If you see the
+      // wrong behavior it is due to the above IF being evaluated incorrectly.
       // When comparing the `customColorSwatchesAndNames` new and old Widget
-      // maps toString results, it was observed they they were equal for the
+      // maps toString results, it was observed that they were equal for the
       // problem use cases, while the [mapEquals] or == operator was not.
       // Therefore using `toString` comparisons for now to get around the issue,
-      // not ideal, but it seems to work.
-
+      // not ideal, but it works OK.
       if (_debug) {
         debugPrint('didUpdateWidget pickersEnabled or custom swatch updated!');
       }
@@ -1339,9 +1538,14 @@ class _ColorPickerState extends State<ColorPicker> {
         ColorPickerType.bw: ColorTools.blackAndWhite,
         ColorPickerType.custom: widget.customColorSwatchesAndNames.keys
             .toList(), // Use empty map if no custom swatch given.
+        ColorPickerType.customSecondary: widget
+            .customSecondaryColorSwatchesAndNames.keys
+            .toList(), // Use empty map if no custom swatch given.
         ColorPickerType.wheel: <ColorSwatch<Object>>[
           // Make a swatch of the selected color in the wheel.
-          ColorTools.primarySwatch(_selectedColor.withAlpha(0xFF))
+          // If color has opacity, it will make a swatch of the color with
+          // same opacity in it as well.
+          ColorTools.primarySwatch(_selectedColor)
         ],
       };
       // Update enabled color pickers, with defaults if none given, depends
@@ -1359,6 +1563,10 @@ class _ColorPickerState extends State<ColorPicker> {
             // Custom picker is always disabled if no custom swatches are given.
             (widget.pickersEnabled[ColorPickerType.custom] ?? false) &&
                 widget.customColorSwatchesAndNames.isNotEmpty,
+        ColorPickerType.customSecondary:
+            // Custom second is always disabled if no custom swatches are given.
+            (widget.pickersEnabled[ColorPickerType.customSecondary] ?? false) &&
+                widget.customSecondaryColorSwatchesAndNames.isNotEmpty,
         ColorPickerType.wheel:
             widget.pickersEnabled[ColorPickerType.wheel] ?? false,
       };
@@ -1377,43 +1585,16 @@ class _ColorPickerState extends State<ColorPicker> {
     if (!listEquals(widget.recentColors, _recentColors)) {
       _recentColors = <Color>[...widget.recentColors];
     }
-    // The widget color was updated externally since it differs from internally
-    // kept state, we should update the widget to new color and find picker.
-    if (widget.color != _selectedColor) {
-      if (_debug) {
-        debugPrint('didUpdateWidget external color update!');
-      }
-      _selectedColor =
-          widget.enableOpacity ? widget.color : widget.color.withAlpha(0xFF);
-      _opacity = widget.enableOpacity ? widget.color.opacity : 1;
-      // Need to make a swatch too be able to find it on wheel, if it is there.
-      _typeToSwatchMap[ColorPickerType.wheel] = <ColorSwatch<Object>>[
-        ColorTools.createPrimarySwatch(_selectedColor.withAlpha(0xFF)),
-      ];
-      // Wheel and edit needs to update.
-      _wheelShouldUpdate = true;
-      _editShouldUpdate = true;
-      // Tonal picker should update from external change.
-      _tonalShouldUpdate = true;
-      _tonalOperated = false;
-      // We need to find the right picker again.
-      shouldFindPickerAndSwatch = true;
-    }
-    //
+    // Last find picker and swatch, if the flag to do so is set.
     if (shouldFindPickerAndSwatch) {
-      if (_debug) {
-        debugPrint('didUpdateWidget calls findPicker and updateActiveSwatch');
-      }
-      // When in this IF branch, we need to find the right picker again.
+      if (_debug) debugPrint('didUpdateWidget shouldFindPickerAndSwatch');
       _findPicker();
-      // And also update the active swatch again.
       _updateActiveSwatch();
     }
-    //
-    super.didUpdateWidget(oldWidget);
+    _fromInternal = false;
   }
 
-  // Find the best matching picker of available ones to show current color.
+  // Find the best matching picker of available ones to show selected color.
   void _findPicker() {
     // The selected color indicator in the picker should request focus.
     _selectedShouldFocus = true;
@@ -1425,13 +1606,11 @@ class _ColorPickerState extends State<ColorPicker> {
     // the same thing! :)
     _usePickerSelector =
         _pickers.values.fold<int>(0, (int t, bool e) => t + (e ? 1 : 0)) > 1;
-
     // If we have a picker selector, we get the best one of the enabled ones,
-    // to show the current _selectedColor.
+    // to show the current selectedColor.
     if (_usePickerSelector) {
       _activePicker = findColorInSelector(
-        // Disregard any alpha on selected color to find its color value only.
-        color: _selectedColor.withAlpha(0xFF),
+        color: _tappedColor,
         typeToSwatchMap: _typeToSwatchMap,
         pickersEnabled: _pickers,
         lookInShades: widget.enableShadesSelection,
@@ -1451,6 +1630,8 @@ class _ColorPickerState extends State<ColorPicker> {
         _activePicker = ColorPickerType.bw;
       } else if (_pickers[ColorPickerType.custom]!) {
         _activePicker = ColorPickerType.custom;
+      } else if (_pickers[ColorPickerType.customSecondary]!) {
+        _activePicker = ColorPickerType.customSecondary;
       } else if (_pickers[ColorPickerType.wheel]!) {
         _activePicker = ColorPickerType.wheel;
       }
@@ -1474,34 +1655,46 @@ class _ColorPickerState extends State<ColorPicker> {
   // Update shades swatch to the correct swatch for the selected color.
   void _updateActiveSwatch() {
     // The typical case is that we have a normal swatch where we need to
-    // find the swatch that contains the _selectedColor.
+    // find the swatch that contains the selectedColor.
     if (_activePicker != ColorPickerType.wheel && !_tonalOperated) {
       // Get list of color swatches from the map for the active picker.
       _activeColorSwatchList = _typeToSwatchMap[_activePicker]!;
+
+      if (_debug) {
+        debugPrint('_updateActiveSwatch _tappedColor= $_tappedColor');
+      }
       // Find the swatch that selected color belongs to from the swatches in
       // the active picker and set this swatch as _activeSwatch.
       _activeSwatch = findColorSwatch(
-        _selectedColor.withAlpha(0xFF),
+        _tappedColor,
         _activeColorSwatchList,
         widget.includeIndex850,
       ) as ColorSwatch<Object>?;
+      if (_debug) {
+        debugPrint('_updateActiveSwatch _activeSwatch= $_activeSwatch');
+      }
       // For the wheel picker we need to check if the selected color belongs to
       // a pre-defined swatch and if it does, return that as the active swatch.
       // If the selected color does not belong to any pre-defined color swatch,
       // then we compute a color swatch for it.
     } else {
-      if (ColorTools.isAccentColor(_selectedColor.withAlpha(0xFF))) {
-        _activeSwatch = ColorTools.accentSwatch(_selectedColor.withAlpha(0xFF));
-      } else if (ColorTools.isPrimaryColor(_selectedColor.withAlpha(0xFF))) {
-        _activeSwatch =
-            ColorTools.primarySwatch(_selectedColor.withAlpha(0xFF));
-      } else if (ColorTools.isBlackAndWhiteColor(
-          _selectedColor.withAlpha(0xFF))) {
-        _activeSwatch =
-            ColorTools.blackAndWhiteSwatch(_selectedColor.withAlpha(0xFF));
+      if (ColorTools.isAccentColor(_tappedColor)) {
+        _activeSwatch = ColorTools.accentSwatch(_tappedColor);
+      } else if (ColorTools.isPrimaryColor(_tappedColor)) {
+        _activeSwatch = ColorTools.primarySwatch(_tappedColor);
+      } else if (ColorTools.isBlackAndWhiteColor(_tappedColor)) {
+        _activeSwatch = ColorTools.blackAndWhiteSwatch(_tappedColor);
+      } else if (ColorTools.isCustomColor(
+          _tappedColor, widget.customColorSwatchesAndNames)) {
+        _activeSwatch = ColorTools.customSwatch(
+            _tappedColor, widget.customColorSwatchesAndNames);
+      } else if (ColorTools.isCustomColor(
+          _tappedColor, widget.customSecondaryColorSwatchesAndNames)) {
+        _activeSwatch = ColorTools.customSwatch(
+            _tappedColor, widget.customSecondaryColorSwatchesAndNames);
       } else {
         _activeSwatch = ColorTools.customSwatch(
-            _selectedColor.withAlpha(0xFF), widget.customColorSwatchesAndNames);
+            _tappedColor, widget.customSecondaryColorSwatchesAndNames);
       }
     }
     // We did not find the selected color in any active swatch list, in that
@@ -1510,7 +1703,8 @@ class _ColorPickerState extends State<ColorPicker> {
     // situation where a selected color was passed to the color picker that was
     // not found in any of the provided swatches in any enabled pickers.
     // If the wheel picker is enabled, then the color will always be found
-    // in it as a last resort.
+    // in it as a last resort, if not we may end up here, using the first
+    // swatch in the list as the active swatch.
     _activeSwatch ??= _activeColorSwatchList[0];
   }
 
@@ -1532,514 +1726,507 @@ class _ColorPickerState extends State<ColorPicker> {
     final TextStyle effectiveGenericNameStyle =
         (widget.colorNameTextStyle ?? Theme.of(context).textTheme.bodyMedium) ??
             const TextStyle();
-
     // Set the default integer code value text style to bodyMedium if not given.
     final TextStyle effectiveCodeStyle =
         (widget.colorCodeTextStyle ?? Theme.of(context).textTheme.bodyMedium) ??
             const TextStyle();
-
     // The logic below is used to determine if we will have a context menu
     // present at all in the Widget tree.
     final bool useContextMenu = widget.copyPasteBehavior.longPressMenu ||
         widget.copyPasteBehavior.secondaryMenu ||
         widget.copyPasteBehavior.secondaryOnDesktopLongOnDevice ||
         widget.copyPasteBehavior.secondaryOnDesktopLongOnDeviceAndWeb;
-
     // Should keyboard listener grab focus? If neither copy and paste keyboard
     // shortcuts are enabled, there is no need to autofocus, so let's skip it
     // then too, regardless of autofocus setting.
     final bool autoFocus = widget.copyPasteBehavior.autoFocus &&
         (widget.copyPasteBehavior.ctrlC || widget.copyPasteBehavior.ctrlV);
-
-    // We start with a RawKeyboardListener that is used to handle keyboard
-    // copy and paste events.
-    return RawKeyboardListener(
+    // Use a copy paste handler to handle copy and paste keyboard shortcuts,
+    // and also to handle the context menu for copy and paste.
+    return CopyPasteHandler(
+      pasteFromClipboard: _getClipboard,
+      copyToClipboard: _setClipboard,
+      useContextMenu: useContextMenu,
+      useLongPress: widget.copyPasteBehavior.longPressMenu,
+      useSecondaryTapDown: widget.copyPasteBehavior.secondaryMenu,
+      useSecondaryOnDesktopLongOnDevice:
+          widget.copyPasteBehavior.secondaryOnDesktopLongOnDevice,
+      useSecondaryOnDesktopLongOnDeviceAndWeb:
+          widget.copyPasteBehavior.secondaryOnDesktopLongOnDeviceAndWeb,
+      onCopyPasteMenuOpened: () {
+        // If we were on the wheel when the menu got opened, it's
+        // operation got cancelled by the context menu and we need to
+        // call onColorChangeEnd event with the selected color.
+        if (_onWheel) {
+          widget.onColorChangeEnd?.call(_selectedColor);
+          // We set onWheel to false as well, as we are no longer on
+          // the wheel and we handled the event.
+          setState(() {
+            _onWheel = false;
+          });
+        }
+      },
       focusNode: _focusNode,
-      onKey: _handleKeyEvent,
-      autofocus: autoFocus,
-      // If the Copy-Paste context menu feature is enabled we wrap the
-      // entire color picker with a Copy-Paste context menu, if it is not
-      // enabled, it is not a part of the widget tree at all.
-      child: IfWrapper(
-        condition: useContextMenu,
-        builder: (BuildContext context, Widget child) {
-          return ContextCopyPasteMenu(
-            useLongPress: widget.copyPasteBehavior.longPressMenu,
-            useSecondaryTapDown: widget.copyPasteBehavior.secondaryMenu,
-            useSecondaryOnDesktopLongOnDevice:
-                widget.copyPasteBehavior.secondaryOnDesktopLongOnDevice,
-            useSecondaryOnDesktopLongOnDeviceAndWeb:
-                widget.copyPasteBehavior.secondaryOnDesktopLongOnDeviceAndWeb,
-            onSelected: (CopyPasteCommands? value) {
-              if (value == CopyPasteCommands.copy) unawaited(_setClipboard());
-              if (value == CopyPasteCommands.paste) unawaited(_getClipboard());
-            },
-            onOpen: () {
-              // If we were on the wheel when the menu got opened, it's
-              // operation got cancelled by the context menu and we need to
-              // call onColorChangeEnd event with the selected color.
-              if (_onWheel) {
-                // But only if there is call back for it.
-                widget.onColorChangeEnd?.call(_selectedColor);
-                // We set onWheel to false as well, as we are no longer on
-                // the wheel and we handled the event.
-                setState(() {
-                  _onWheel = false;
-                });
-              }
-            },
-            child: child,
-          );
-        },
-        child: Padding(
-          padding: widget.padding,
-          // The ColorPicker layout is a column. It is up to the user to ensure
-          // that it fits, or use a scrolling parent if it does not.
-          child: Column(
-            crossAxisAlignment: widget.crossAxisAlignment,
-            children: <Widget>[
-              // Show title bar widget if we have one.
-              if (widget.title != null ||
-                  widget.copyPasteBehavior.copyButton ||
-                  widget.copyPasteBehavior.pasteButton ||
-                  widget.actionButtons.okButton ||
-                  widget.actionButtons.closeButton)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: ColorPickerToolbar(
-                    title: widget.title,
-                    onCopy: widget.copyPasteBehavior.copyButton
-                        ? _setClipboard
-                        : null,
-                    onPaste: widget.copyPasteBehavior.pasteButton
-                        ? _getClipboard
-                        : null,
-                    onOk: widget.actionButtons.okButton
-                        ? () {
-                            // OK was pressed, we pop and return TRUE.
-                            // In case this was not used in a dialog the
-                            // canPop will at least avoid a crash, but may
-                            // still do the wrong thing.
-                            if (Navigator.of(context,
+      autoFocus: autoFocus,
+      child: Padding(
+        padding: widget.padding,
+        child: Column(
+          crossAxisAlignment: widget.crossAxisAlignment,
+          children: <Widget>[
+            // Show title bar widget if we have one.
+            if (widget.title != null ||
+                widget.copyPasteBehavior.copyButton ||
+                widget.copyPasteBehavior.pasteButton ||
+                widget.actionButtons.okButton ||
+                widget.actionButtons.closeButton)
+              Padding(
+                padding: EdgeInsets.only(
+                    bottom: widget.toolbarSpacing ?? widget.columnSpacing),
+                child: ColorPickerToolbar(
+                  title: widget.title,
+                  onCopy: widget.copyPasteBehavior.copyButton
+                      ? _setClipboard
+                      : null,
+                  onPaste: widget.copyPasteBehavior.pasteButton
+                      ? _getClipboard
+                      : null,
+                  onOk: widget.actionButtons.okButton
+                      ? () {
+                          // OK was pressed, we pop and return TRUE.
+                          // In case this was not used in a dialog the
+                          // canPop will at least avoid a crash, but may
+                          // still do the wrong thing.
+                          if (Navigator.of(context,
+                                  rootNavigator:
+                                      widget.actionButtons.useRootNavigator)
+                              .canPop()) {
+                            Navigator.of(context,
                                     rootNavigator:
                                         widget.actionButtons.useRootNavigator)
-                                .canPop()) {
-                              Navigator.of(context,
-                                      rootNavigator:
-                                          widget.actionButtons.useRootNavigator)
-                                  .pop(true);
-                            }
+                                .pop(true);
                           }
-                        : null,
-                    onClose: widget.actionButtons.closeButton
-                        ? () {
-                            // Cancel was pressed, we pop and return FALSE.
-                            // In case this was not used in a dialog the
-                            // canPop will at least avoid a crash, but may
-                            // still do the wrong thing.
-                            if (Navigator.of(context,
-                                    rootNavigator:
-                                        widget.actionButtons.useRootNavigator)
-                                .canPop()) {
-                              Navigator.of(context,
-                                      rootNavigator:
-                                          widget.actionButtons.useRootNavigator)
-                                  .pop(false);
-                            }
-                          }
-                        : null,
-                    toolIcons: widget.actionButtons,
-                    copyPasteBehavior: widget.copyPasteBehavior,
-                    enableTooltips: widget.enableTooltips,
-                  ),
-                ),
-              // Show heading widget if we have one.
-              if (widget.heading != null)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: widget.heading,
-                ),
-              // Show picker selector, if more than one picker is enabled.
-              if (_usePickerSelector)
-                Focus(
-                  focusNode: _pickerFocusNode,
-                  onKey: (FocusNode node, RawKeyEvent event) {
-                    _handleKeyEvent(event);
-                    return KeyEventResult.ignored;
-                  },
-                  child: SelectPicker(
-                    pickers: _pickers,
-                    pickerLabels: _pickerLabels,
-                    picker: _activePicker,
-                    onPickerChanged: (ColorPickerType value) {
-                      // If there is no color in the picker that will grab
-                      // focus when we move to the picker, then the picker
-                      // selector itself will keep focus, and kick the key
-                      // events (copy/paste) to the handler, we can thus use
-                      // CTRL-C/V also when picker selector has focus.
-                      _pickerFocusNode.requestFocus();
-                      setState(() {
-                        _activePicker = value;
-                        // If there is a color indicator, it will grab focus.
-                        _selectedShouldFocus = true;
-                        // If we are on the wheel no shades selection is not
-                        // enabled, then the wheel will grab the focus.
-                        if (_activePicker == ColorPickerType.wheel &&
-                            !widget.enableShadesSelection &&
-                            !widget.enableTonalPalette) {
-                          _wheelShouldFocus = true;
                         }
-                        _tonalOperated = false;
-                        _updateActiveSwatch();
-                      });
-                    },
-                    thumbColor: widget.selectedPickerTypeColor,
-                    textStyle: widget.pickerTypeTextStyle,
-                    columnSpacing: widget.columnSpacing,
-                  ),
+                      : null,
+                  onClose: widget.actionButtons.closeButton
+                      ? () {
+                          // Cancel was pressed, we pop and return FALSE.
+                          // In case this was not used in a dialog the
+                          // canPop will at least avoid a crash, but may
+                          // still do the wrong thing.
+                          if (Navigator.of(context,
+                                  rootNavigator:
+                                      widget.actionButtons.useRootNavigator)
+                              .canPop()) {
+                            Navigator.of(context,
+                                    rootNavigator:
+                                        widget.actionButtons.useRootNavigator)
+                                .pop(false);
+                          }
+                        }
+                      : null,
+                  toolIcons: widget.actionButtons,
+                  copyPasteBehavior: widget.copyPasteBehavior,
+                  enableTooltips: widget.enableTooltips,
                 ),
-              // Add a tiny bit of extra hard coded space after the picker
-              // type selector if there was one.
-              if (_usePickerSelector) const SizedBox(height: 4),
-              // This is not the Wheel case, so we draw all the main colors
-              // for the active swatch list.
-              if (_activePicker != ColorPickerType.wheel)
-                MainColors(
-                  spacing: widget.spacing,
-                  runSpacing: widget.runSpacing,
-                  columnSpacing: widget.columnSpacing,
-                  activeColorSwatchList: _activeColorSwatchList,
-                  selectedColor: _selectedColor.withAlpha(0xFF),
-                  onSelectColor: (Color color) {
-                    _tonalOperated = false;
-                    _onSelectColor(color);
-                  },
-                  includeIndex850: widget.includeIndex850,
-                  width: widget.width,
-                  height: widget.height,
-                  borderRadius: widget.borderRadius,
-                  hasBorder: widget.hasBorder,
-                  borderColor: widget.borderColor,
-                  elevation: widget.elevation,
-                  selectedColorIcon: widget.selectedColorIcon,
-                  selectedRequestsFocus: _selectedShouldFocus,
-                ),
-              // This is the wheel case, draw the custom ColorWheelPicker.
-              if (_activePicker == ColorPickerType.wheel)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: SizedBox(
-                    height: widget.wheelDiameter,
-                    width: widget.wheelDiameter,
-                    child: ColorWheelPicker(
-                      color: _selectedColor.withAlpha(0xFF),
-                      wheelWidth: widget.wheelWidth,
-                      wheelSquarePadding: widget.wheelSquarePadding,
-                      wheelSquareBorderRadius: widget.wheelSquareBorderRadius,
-                      hasBorder: widget.wheelHasBorder,
-                      borderColor: widget.borderColor,
-                      shouldUpdate: _wheelShouldUpdate,
-                      shouldRequestsFocus: _wheelShouldFocus,
-                      onChangeStart: (Color color) {
-                        widget.onColorChangeStart
-                            ?.call(color.withOpacity(_opacity));
-                        _addToRecentColors(color.withOpacity(_opacity));
-                      },
-                      onChanged: (Color color) {
-                        setState(() {
-                          _selectedColor = color.withOpacity(_opacity);
-                          _wheelShouldUpdate = false;
-                          _editShouldUpdate = true;
-                          _tonalShouldUpdate = true;
-                          _tonalOperated = false;
-                          _selectedShouldFocus = true;
-                          _wheelShouldFocus = false;
-                          _updateActiveSwatch();
-                        });
-                        widget.onColorChanged(_selectedColor);
-                      },
-                      onChangeEnd: (Color color) {
-                        widget.onColorChangeEnd?.call(
-                          color.withOpacity(_opacity),
-                        );
-                      },
-                      onWheel: (bool value) {
-                        setState(() {
-                          _onWheel = value;
-                        });
-                      },
-                    ),
-                  ),
-                ),
-              // Show the sub-heading for the none wheel case.
-              if (widget.subheading != null &&
-                  widget.enableShadesSelection &&
-                  _activePicker != ColorPickerType.wheel)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: widget.subheading,
-                ),
-              // Show the sub-heading for the wheel case.
-              if (widget.wheelSubheading != null &&
-                  widget.enableShadesSelection &&
-                  _activePicker == ColorPickerType.wheel)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: widget.wheelSubheading,
-                ),
-              // Draw the shade colors for the selected main color.
-              if (widget.enableShadesSelection)
-                ShadeColors(
-                  spacing: widget.spacing,
-                  runSpacing: widget.runSpacing,
-                  columnSpacing: widget.columnSpacing,
-                  activeSwatch: _activeSwatch!,
-                  selectedColor: _selectedColor.withAlpha(0xFF),
-                  onSelectColor: (Color color) {
-                    _tonalOperated = false;
-                    _onSelectColor(color);
-                    if (_activePicker == ColorPickerType.wheel) {
-                      setState(() {
-                        _selectedShouldFocus = true;
-                        _tonalShouldUpdate = true;
-                      });
-                    }
-                  },
-                  includeIndex850: widget.includeIndex850,
-                  width: widget.width,
-                  height: widget.height,
-                  borderRadius: widget.borderRadius,
-                  hasBorder: widget.hasBorder,
-                  borderColor: widget.borderColor,
-                  elevation: widget.elevation,
-                  selectedColorIcon: widget.selectedColorIcon,
-                  selectedRequestsFocus: _selectedShouldFocus,
-                ),
-              // Show the tonal sub-heading
-              if (widget.tonalSubheading != null && widget.enableTonalPalette)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: widget.tonalSubheading,
-                ),
-              // Draw the tonal palette.
-              if (widget.enableTonalPalette)
-                TonalPaletteColors(
-                  spacing: widget.spacing,
-                  runSpacing: widget.runSpacing,
-                  columnSpacing: widget.columnSpacing,
-                  selectedColor: _selectedColor.withAlpha(0xFF),
-                  onSelectColor: (Color color) {
-                    _tonalOperated = true;
-                    _onSelectColor(color);
+              ),
+            // Show heading widget if we have one.
+            if (widget.heading != null)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: widget.heading,
+              ),
+            // Show picker selector, if more than one picker is enabled.
+            if (_usePickerSelector)
+              Focus(
+                focusNode: _pickerFocusNode,
+                child: SelectPicker(
+                  pickers: _pickers,
+                  pickerLabels: _pickerLabels,
+                  picker: _activePicker,
+                  onPickerChanged: (ColorPickerType value) {
+                    // If there is no color in the picker that will grab
+                    // focus when we move to the picker, then the picker
+                    // selector itself will keep focus, and kick the key
+                    // events (copy/paste) to the handler, we can thus use
+                    // CTRL-C/V also when picker selector has focus.
+                    _pickerFocusNode.requestFocus();
                     setState(() {
+                      _activePicker = value;
+                      // If there is a color indicator, it will grab focus.
                       _selectedShouldFocus = true;
-                      _tonalShouldUpdate = false;
+                      // If we are on the wheel no shades selection is not
+                      // enabled, then the wheel will grab the focus.
+                      if (_activePicker == ColorPickerType.wheel &&
+                          !widget.enableShadesSelection &&
+                          !widget.enableTonalPalette) {
+                        _wheelShouldFocus = true;
+                      }
+                      _tonalOperated = false;
+                      _updateActiveSwatch();
                     });
                   },
-                  tonalShouldUpdate: _tonalShouldUpdate,
-                  width: widget.tonalColorSameSize
-                      ? widget.width
-                      : (widget.width + widget.spacing) * 10 / 15 -
-                          widget.spacing,
-                  height: widget.tonalColorSameSize
-                      ? widget.width
-                      : (widget.width + widget.spacing) * 10 / 15 -
-                          widget.spacing,
-                  borderRadius: widget.borderRadius,
-                  hasBorder: widget.hasBorder,
-                  borderColor: widget.borderColor,
-                  elevation: widget.elevation,
-                  selectedColorIcon: widget.selectedColorIcon,
-                  selectedRequestsFocus: _selectedShouldFocus,
+                  thumbColor: widget.selectedPickerTypeColor,
+                  textStyle: widget.pickerTypeTextStyle,
+                  columnSpacing: widget.columnSpacing,
                 ),
-              // Show the sub-heading for the opacity control.
-              if (widget.opacitySubheading != null && widget.enableOpacity)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: widget.opacitySubheading,
+              ),
+            // Add a tiny bit of extra hard coded space after the picker
+            // type selector if there was one.
+            if (_usePickerSelector) const SizedBox(height: 4),
+            // This is not the Wheel case, so we draw all the main colors
+            // for the active swatch list.
+            if (_activePicker != ColorPickerType.wheel)
+              MainColors(
+                spacing: widget.spacing,
+                runSpacing: widget.runSpacing,
+                columnSpacing: widget.columnSpacing,
+                activeColorSwatchList: _activeColorSwatchList,
+                selectedColor:
+                    widget.enableOpacity ? _selectedColor : _tappedColor,
+                onSelectColor: (Color color) {
+                  _tonalOperated = false;
+                  _onSelectColor(color);
+                },
+                includeIndex850: widget.includeIndex850,
+                width: widget.width,
+                height: widget.height,
+                borderRadius: widget.borderRadius,
+                hasBorder: widget.hasBorder,
+                borderColor: widget.borderColor,
+                elevation: widget.elevation,
+                selectedColorIcon: widget.selectedColorIcon,
+                selectedRequestsFocus: _selectedShouldFocus,
+              ),
+            // This is the wheel case, draw the custom ColorWheelPicker.
+            if (_activePicker == ColorPickerType.wheel)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: SizedBox(
+                  height: widget.wheelDiameter,
+                  width: widget.wheelDiameter,
+                  child: ColorWheelPicker(
+                    color: _selectedColor,
+                    wheelWidth: widget.wheelWidth,
+                    wheelSquarePadding: widget.wheelSquarePadding,
+                    wheelSquareBorderRadius: widget.wheelSquareBorderRadius,
+                    hasBorder: widget.wheelHasBorder,
+                    borderColor: widget.borderColor,
+                    shouldUpdate: _wheelShouldUpdate,
+                    shouldRequestsFocus: _wheelShouldFocus,
+                    onChangeStart: (Color color) {
+                      widget.onColorChangeStart
+                          ?.call(color.withOpacity(_opacity));
+                      _addToRecentColors(color.withOpacity(_opacity));
+                    },
+                    onChanged: (Color color) {
+                      setState(() {
+                        _tappedColor = color;
+                        _selectedColor = color;
+                        _wheelShouldUpdate = false;
+                        _editShouldUpdate = true;
+                        _tonalShouldUpdate = true;
+                        _tonalOperated = false;
+                        _selectedShouldFocus = true;
+                        _wheelShouldFocus = false;
+                        _updateActiveSwatch();
+                      });
+                      widget
+                          .onColorChanged(_selectedColor.withOpacity(_opacity));
+                    },
+                    onChangeEnd: (Color color) {
+                      widget.onColorChangeEnd?.call(
+                        color.withOpacity(_opacity),
+                      );
+                    },
+                    onWheel: (bool value) {
+                      setState(() {
+                        _onWheel = value;
+                      });
+                    },
+                  ),
                 ),
-              // Draw the opacity slider if enabled.
-              if (widget.enableOpacity)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: SizedBox(
-                    width: (widget.opacityTrackWidth ?? 0) < 150
-                        ? double.infinity
-                        : widget.opacityTrackWidth,
-                    child: RepaintBoundary(
-                      child: OpacitySlider(
-                        color: _selectedColor.withAlpha(0xFF),
-                        opacity: _opacity,
-                        trackHeight: widget.opacityTrackHeight,
-                        thumbRadius: widget.opacityThumbRadius,
-                        focusNode: _opacityFocusNode,
-                        onChangeStart: (double value) {
-                          if (widget.onColorChangeStart != null) {
-                            // Request focus when change starts.
-                            _opacityFocusNode.requestFocus();
-                            setState(() {
-                              _opacity = value;
-                              _selectedColor =
-                                  _selectedColor.withOpacity(_opacity);
-                            });
-                            widget.onColorChangeStart!(_selectedColor);
-                          }
-                          _addToRecentColors(_selectedColor);
-                        },
-                        onChanged: (double value) {
+              ),
+            // Show the sub-heading for the none wheel case.
+            if (widget.subheading != null &&
+                widget.enableShadesSelection &&
+                _activePicker != ColorPickerType.wheel)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: widget.subheading,
+              ),
+            // Show the sub-heading for the wheel case.
+            if (widget.wheelSubheading != null &&
+                widget.enableShadesSelection &&
+                _activePicker == ColorPickerType.wheel)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: widget.wheelSubheading,
+              ),
+            // Draw the shade colors for the selected main color.
+            if (widget.enableShadesSelection)
+              ShadeColors(
+                spacing: widget.spacing,
+                runSpacing: widget.runSpacing,
+                columnSpacing: widget.shadesSpacing ?? widget.columnSpacing,
+                activeSwatch: _activeSwatch!,
+                selectedColor:
+                    widget.enableOpacity ? _selectedColor : _tappedColor,
+                onSelectColor: (Color color) {
+                  _tonalOperated = false;
+                  _onSelectColor(color);
+                  if (_activePicker == ColorPickerType.wheel) {
+                    setState(() {
+                      _selectedShouldFocus = true;
+                      _tonalShouldUpdate = true;
+                    });
+                  }
+                },
+                includeIndex850: widget.includeIndex850,
+                width: widget.width,
+                height: widget.height,
+                borderRadius: widget.borderRadius,
+                hasBorder: widget.hasBorder,
+                borderColor: widget.borderColor,
+                elevation: widget.elevation,
+                selectedColorIcon: widget.selectedColorIcon,
+                selectedRequestsFocus: _selectedShouldFocus,
+              ),
+            // Show the tonal sub-heading
+            if (widget.tonalSubheading != null && widget.enableTonalPalette)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: widget.tonalSubheading,
+              ),
+            // Draw the tonal palette.
+            if (widget.enableTonalPalette) ...<Widget>[
+              TonalPaletteColors(
+                spacing: widget.spacing,
+                runSpacing: widget.runSpacing,
+                selectedColor: _tappedColor,
+                onSelectColor: (Color color) {
+                  _tonalOperated = true;
+                  _onSelectColor(color);
+                  setState(() {
+                    _selectedShouldFocus = true;
+                    _tonalShouldUpdate = false;
+                  });
+                },
+                tonalShouldUpdate: _tonalShouldUpdate,
+                width: widget.tonalColorSameSize
+                    ? widget.width
+                    : (widget.width + widget.spacing) * 10 / 15 -
+                        widget.spacing,
+                height: widget.tonalColorSameSize
+                    ? widget.width
+                    : (widget.width + widget.spacing) * 10 / 15 -
+                        widget.spacing,
+                borderRadius: widget.borderRadius,
+                hasBorder: widget.hasBorder,
+                borderColor: widget.borderColor,
+                elevation: widget.elevation,
+                selectedColorIcon: widget.selectedColorIcon,
+                selectedRequestsFocus: _selectedShouldFocus,
+              ),
+              SizedBox(height: widget.columnSpacing),
+            ],
+            // Show the sub-heading for the opacity control.
+            if (widget.opacitySubheading != null && widget.enableOpacity)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: widget.opacitySubheading,
+              ),
+            // Draw the opacity slider if enabled.
+            if (widget.enableOpacity)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: SizedBox(
+                  width: (widget.opacityTrackWidth ?? 0) < 150
+                      ? double.infinity
+                      : widget.opacityTrackWidth,
+                  child: RepaintBoundary(
+                    child: OpacitySlider(
+                      color: _selectedColor.withAlpha(0xFF),
+                      opacity: _opacity,
+                      trackHeight: widget.opacityTrackHeight,
+                      thumbRadius: widget.opacityThumbRadius,
+                      focusNode: _opacityFocusNode,
+                      onChangeStart: (double value) {
+                        if (widget.onColorChangeStart != null) {
+                          // Request focus when change starts.
+                          _opacityFocusNode.requestFocus();
                           setState(() {
+                            _fromInternal = true;
                             _opacity = value;
-                            _selectedColor =
-                                _selectedColor.withOpacity(_opacity);
-                            _wheelShouldUpdate = false;
-                            _editShouldUpdate = true;
-                            _selectedShouldFocus = true;
-                            _wheelShouldFocus = false;
                           });
-                          widget.onColorChanged(_selectedColor);
-                        },
-                        onChangeEnd: (double value) {
-                          if (widget.onColorChangeEnd != null) {
-                            setState(() {
-                              _opacity = value;
-                              _selectedColor =
-                                  _selectedColor.withOpacity(_opacity);
-                            });
-                            widget.onColorChangeEnd!(_selectedColor);
-                          }
-                        },
-                      ),
+                          widget.onColorChangeStart!(
+                              _selectedColor.withOpacity(_opacity));
+                          _addToRecentColors(
+                              _selectedColor.withOpacity(_opacity));
+                        }
+                      },
+                      onChanged: (double value) {
+                        setState(() {
+                          _fromInternal = true;
+                          _opacity = value;
+                          _wheelShouldUpdate = false;
+                          _editShouldUpdate = true;
+                          _selectedShouldFocus = true;
+                          _wheelShouldFocus = false;
+                        });
+                        widget.onColorChanged(
+                            _selectedColor.withOpacity(_opacity));
+                      },
+                      onChangeEnd: (double value) {
+                        if (widget.onColorChangeEnd != null) {
+                          setState(() {
+                            _fromInternal = true;
+                            _opacity = value;
+                          });
+                          widget.onColorChangeEnd!(
+                              _selectedColor.withOpacity(_opacity));
+                          // _addToRecentColors(
+                          //     _selectedColor.withOpacity(_opacity));
+                        }
+                      },
                     ),
                   ),
                 ),
-              // If we show material or generic name, we enclose them in a
-              // Wrap, they will be on same row nicely if there is room
-              // enough, but also wrap to two rows when so needed when both
-              // are shown at the same and they don't fit on one row.
-              if (widget.showMaterialName || widget.showColorName)
-                Wrap(
-                  children: <Widget>[
-                    // Show the Material color name, if enabled.
-                    if (widget.showMaterialName)
-                      Padding(
-                        padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                        child: Text(
-                          ColorTools.materialName(
-                            _selectedColor.withAlpha(0xFF),
-                            colorSwatchNameMap:
-                                widget.customColorSwatchesAndNames,
-                          ),
-                          style: effectiveMaterialNameStyle,
+              ),
+            // If we show material or generic name, we enclose them in a
+            // Wrap, they will be on same row nicely if there is room
+            // enough, but also wrap to two rows when so needed when both
+            // are shown at the same and they don't fit on one row.
+            if (widget.showMaterialName || widget.showColorName)
+              Wrap(
+                children: <Widget>[
+                  // Show the Material color name, if enabled.
+                  if (widget.showMaterialName)
+                    Padding(
+                      padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                      child: Text(
+                        ColorTools.materialName(
+                          _selectedColor.withAlpha(0xFF),
+                          colorSwatchNameMap:
+                              widget.customColorSwatchesAndNames,
                         ),
+                        style: effectiveMaterialNameStyle,
                       ),
-                    // If we show both material and generic name, add some
-                    // hard coded horizontal space between them.
-                    if (widget.showMaterialName && widget.showColorName)
+                    ),
+                  // If we show both material and generic name, add some
+                  // hard coded horizontal space between them.
+                  if (widget.showMaterialName && widget.showColorName)
+                    const SizedBox(width: 8),
+                  // Show the generic color name, if enabled.
+                  if (widget.showColorName)
+                    Padding(
+                      padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                      child: Text(
+                        ColorTools.nameThatColor(
+                            _selectedColor.withAlpha(0xFF)),
+                        style: effectiveGenericNameStyle,
+                      ),
+                    ),
+                ],
+              ),
+            // If we show color code or its int value, we enclose them in a
+            // Wrap, they will be on same row nicely if there is room enough
+            // but also wrap to two rows when so needed when both are
+            // shown at the same and they don't fit on one row.
+            if (widget.showColorCode || widget.showColorValue)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: Wrap(
+                  crossAxisAlignment: WrapCrossAlignment.center,
+                  runAlignment: WrapAlignment.center,
+                  alignment: WrapAlignment.center,
+                  children: <Widget>[
+                    // Show the color code view and edit field, if enabled.
+                    if (widget.showColorCode)
+                      ColorCodeField(
+                        color: widget.enableOpacity
+                            ? _selectedColor.withOpacity(_opacity)
+                            : _tappedColor,
+                        readOnly: _activePicker != ColorPickerType.wheel ||
+                            widget.colorCodeReadOnly,
+                        textStyle: widget.colorCodeTextStyle,
+                        prefixStyle: widget.colorCodePrefixStyle,
+                        colorCodeHasColor: widget.colorCodeHasColor,
+                        enableTooltips: widget.enableTooltips,
+                        shouldUpdate: _editShouldUpdate,
+                        onColorChanged: (Color color) {
+                          widget.onColorChangeStart
+                              ?.call(_selectedColor.withOpacity(_opacity));
+                          setState(() {
+                            _tappedColor = color;
+                            _selectedColor = color;
+                            // Color changed outside wheel picker, when the
+                            // code was edited, the wheel should update.
+                            _wheelShouldUpdate = true;
+                            _editShouldUpdate = false;
+                            _tonalOperated = false;
+                            _updateActiveSwatch();
+                          });
+                          widget.onColorChanged(
+                              _selectedColor.withOpacity(_opacity));
+                          widget.onColorChangeEnd
+                              ?.call(_selectedColor.withOpacity(_opacity));
+                          _addToRecentColors(
+                              _selectedColor.withOpacity(_opacity));
+                        },
+                        onEditFocused: (bool editInFocus) {
+                          setState(() {
+                            _editCodeFocused = editInFocus;
+                            if (_editCodeFocused) {
+                              _selectedShouldFocus = false;
+                              _wheelShouldFocus = false;
+                            }
+                          });
+                        },
+                        toolIcons: widget.actionButtons,
+                        copyPasteBehavior: widget.copyPasteBehavior,
+                      ),
+                    // If we show both hex code and int value, add some
+                    // hardcoded horizontal space between them.
+                    if (widget.showColorCode && widget.showColorValue)
                       const SizedBox(width: 8),
-                    // Show the generic color name, if enabled.
-                    if (widget.showColorName)
-                      Padding(
-                        padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                        child: Text(
-                          ColorTools.nameThatColor(
-                              _selectedColor.withAlpha(0xFF)),
-                          style: effectiveGenericNameStyle,
-                        ),
+                    if (widget.showColorValue)
+                      SelectableText(
+                        _selectedColor.value.toString(),
+                        style: effectiveCodeStyle,
                       ),
                   ],
                 ),
-              // If we show color code or its int value, we enclose them in a
-              // Wrap, they will be on same row nicely if there is room enough
-              // but also wrap to two rows when so needed when both are
-              // shown at the same and they don't fit on one row.
-              if (widget.showColorCode || widget.showColorValue)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: Wrap(
-                    crossAxisAlignment: WrapCrossAlignment.center,
-                    runAlignment: WrapAlignment.center,
-                    alignment: WrapAlignment.center,
-                    children: <Widget>[
-                      // Show the color code view and edit field, if enabled.
-                      if (widget.showColorCode)
-                        ColorCodeField(
-                          color: _selectedColor,
-                          readOnly: _activePicker != ColorPickerType.wheel ||
-                              widget.colorCodeReadOnly,
-                          textStyle: widget.colorCodeTextStyle,
-                          prefixStyle: widget.colorCodePrefixStyle,
-                          colorCodeHasColor: widget.colorCodeHasColor,
-                          enableTooltips: widget.enableTooltips,
-                          shouldUpdate: _editShouldUpdate,
-                          onColorChanged: (Color color) {
-                            widget.onColorChangeStart?.call(_selectedColor);
-                            setState(() {
-                              _selectedColor = color.withOpacity(_opacity);
-                              // Color changed outside wheel picker, when the
-                              // code was edited, the wheel should update.
-                              _wheelShouldUpdate = true;
-                              _editShouldUpdate = false;
-                              _tonalOperated = false;
-                              _updateActiveSwatch();
-                            });
-                            widget.onColorChanged(_selectedColor);
-                            widget.onColorChangeEnd?.call(_selectedColor);
-                            _addToRecentColors(color);
-                          },
-                          onEditFocused: (bool editInFocus) {
-                            setState(() {
-                              _editCodeFocused = editInFocus;
-                              if (_editCodeFocused) {
-                                _selectedShouldFocus = false;
-                                _wheelShouldFocus = false;
-                              }
-                            });
-                          },
-                          toolIcons: widget.actionButtons,
-                          copyPasteBehavior: widget.copyPasteBehavior,
-                        ),
-                      // If we show both hex code and int value, add some
-                      // hardcoded horizontal space between them.
-                      if (widget.showColorCode && widget.showColorValue)
-                        const SizedBox(width: 8),
-                      if (widget.showColorValue)
-                        SelectableText(
-                          _selectedColor.value.toString(),
-                          style: effectiveCodeStyle,
-                        ),
-                    ],
-                  ),
-                ),
-              // Show the sub-heading for recent colors.
-              if (widget.recentColorsSubheading != null &&
-                  widget.showRecentColors)
-                Padding(
-                  padding: EdgeInsets.only(bottom: widget.columnSpacing),
-                  child: widget.recentColorsSubheading,
-                ),
-              // Show the color indicators for the recently used colors.
-              if (widget.showRecentColors)
-                RecentColors(
-                  spacing: widget.spacing,
-                  runSpacing: widget.runSpacing,
-                  columnSpacing: widget.columnSpacing,
-                  recentColors: _recentColors,
-                  selectedColor: _selectedColor,
-                  onSelectColor: (Color color) {
-                    _onSelectColor(color, keepOpacity: true, findPicker: true);
-                  },
-                  includeIndex850: widget.includeIndex850,
-                  width: widget.width,
-                  height: widget.height,
-                  borderRadius: widget.borderRadius,
-                  hasBorder: widget.hasBorder,
-                  borderColor: widget.borderColor,
-                  elevation: widget.elevation,
-                  selectedColorIcon: widget.selectedColorIcon,
-                  selectedRequestsFocus: _selectedShouldFocus,
-                ),
-            ],
-          ),
+              ),
+            // Show the sub-heading for recent colors.
+            if (widget.recentColorsSubheading != null &&
+                widget.showRecentColors)
+              Padding(
+                padding: EdgeInsets.only(bottom: widget.columnSpacing),
+                child: widget.recentColorsSubheading,
+              ),
+            // Show the color indicators for the recently used colors.
+            if (widget.showRecentColors)
+              RecentColors(
+                spacing: widget.spacing,
+                runSpacing: widget.runSpacing,
+                recentColors: _recentColors,
+                selectedColor: _tappedColor,
+                onSelectColor: (Color color) {
+                  _tonalOperated = false;
+                  _onSelectColor(color, findPicker: true);
+                },
+                includeIndex850: widget.includeIndex850,
+                width: widget.width,
+                height: widget.height,
+                borderRadius: widget.borderRadius,
+                hasBorder: widget.hasBorder,
+                borderColor: widget.borderColor,
+                elevation: widget.elevation,
+                selectedColorIcon: widget.selectedColorIcon,
+                selectedRequestsFocus: _selectedShouldFocus,
+              ),
+            if (widget.showRecentColors) SizedBox(height: widget.columnSpacing),
+          ],
         ),
       ),
     );
@@ -2048,25 +2235,37 @@ class _ColorPickerState extends State<ColorPicker> {
   // A color was selected by tapping it, update state and notify via callbacks.
   void _onSelectColor(
     Color color, {
-    // Set to true to use opacity in `color` and let it replace _opacity.
-    // Done when selecting colors in recently used colors, otherwise not.
-    bool keepOpacity = false,
     // Normally when colors are selected, we do not need find the picker as they
     // are in the same picker. However, recently used colors sets this to true
     // as its colors can be in any picker, so it must be found.
     bool findPicker = false,
   }) {
+    // Any callbacks that are called here, should set the _fromInternal flag.
+    _fromInternal = true;
+
+    // If findPicker is true we came from RecentColors, it is the only one that
+    // uses this flag. If we use Recent colors, we should also extract the
+    // opacity from the selected color, and set it as the current opacity.
+    if (findPicker) {
+      _opacity = color.opacity;
+    }
     // Call start callback with current selectedColor before change.
-    widget.onColorChangeStart?.call(_selectedColor);
-    // Add the previously selected color to recent colors.
-    _addToRecentColors(_selectedColor);
+    if (widget.enableOpacity) {
+      widget.onColorChangeStart?.call(_selectedColor.withOpacity(_opacity));
+      _addToRecentColors(_selectedColor.withOpacity(_opacity));
+    } else {
+      // This is to allow custom colors with opacity to callback with their
+      // in-built opacity, when opacity is control is not enabled.
+      widget.onColorChangeStart?.call(_tappedColor);
+      _addToRecentColors(_tappedColor);
+    }
     // update the state of the selectedColor to the new selected color.
     setState(() {
-      // Set selected color to the new value.
-      _selectedColor = keepOpacity ? color : color.withOpacity(_opacity);
-      if (keepOpacity && widget.enableOpacity) {
-        _opacity = _selectedColor.opacity;
-      }
+      // Set selected color to the one that was "clicked"
+      _tappedColor = color;
+      _selectedColor = _tappedColor.withAlpha(0xFF);
+      if (_debug) debugPrint('_onSelectColor _tappedColor: $_tappedColor');
+      if (_debug) debugPrint('_onSelectColor _selectedColor: $_selectedColor');
       // When the a color was clicked and selected, the right item is already
       // focused an other selected color indicators and wheel should not focus.
       _selectedShouldFocus = false;
@@ -2082,7 +2281,7 @@ class _ColorPickerState extends State<ColorPicker> {
       if (findPicker) {
         // Make a swatch of the selected color in the wheel.
         _typeToSwatchMap[ColorPickerType.wheel] = <ColorSwatch<Object>>[
-          ColorTools.createPrimarySwatch(_selectedColor.withAlpha(0xFF))
+          ColorTools.createPrimarySwatch(_selectedColor)
         ];
         _findPicker();
       }
@@ -2090,18 +2289,29 @@ class _ColorPickerState extends State<ColorPicker> {
       // of the current selected swatch, then update the active swatch.
       // This check eliminates a rebuild when the selected color is a member
       // of the currently displayed color swatch.
+      // Needed to update the active swatch when opacity is used on custom
+      // colors when opacity is not enabled.
       if ((_activePicker != ColorPickerType.wheel ||
               !ColorTools.swatchContainsColor(
                   _activeSwatch!, _selectedColor)) &&
           !_tonalOperated) {
         // Update the active swatch to match the selected color.
+        if (_debug) {
+          debugPrint('**** _onSelectColor calls _updateActiveSwatch ****');
+        }
         _updateActiveSwatch();
       }
     });
     // Call the change call back with the new color.
-    widget.onColorChanged(_selectedColor);
-    // We have and end callback, call it with the new color.
-    widget.onColorChangeEnd?.call(_selectedColor);
+    if (widget.enableOpacity) {
+      widget.onColorChanged(_selectedColor.withOpacity(_opacity));
+      widget.onColorChangeEnd?.call(_selectedColor.withOpacity(_opacity));
+    } else {
+      // This is to allow custom colors with opacity to callback with their
+      // in-built opacity, when opacity is control is not enabled.
+      widget.onColorChanged(_tappedColor);
+      widget.onColorChangeEnd?.call(_tappedColor);
+    }
   }
 
   void _addToRecentColors(Color color) {
@@ -2121,67 +2331,6 @@ class _ColorPickerState extends State<ColorPicker> {
     });
     // Call callback for the handling recent colors, if there is one.
     widget.onRecentColorsChanged?.call(_recentColors);
-  }
-
-  // Handle the keyboard events from the RawKeyboardListener.
-  void _handleKeyEvent(RawKeyEvent event) {
-    // If edit color code is focused and we do not use the parsed paste feature,
-    // we exit out of here. The TextField's normal paste action will then handle
-    // the paste as before in v1.x and normally in a TextField.
-    if (_editCodeFocused && !widget.copyPasteBehavior.editUsesParsedPaste) {
-      return;
-    }
-    // Make an OS independent copy/paste shortcut modifier key.
-    //
-    // Found the usage of these in the SDK TextField copy/paste implementation.
-    final bool isRawKeyMacOS = event.data is RawKeyEventDataMacOs;
-    if (_debug) {
-      debugPrint('KeyEvent isRawKeyMacOS: $isRawKeyMacOS');
-    }
-    final bool isRawKeyIos = event.data is RawKeyEventDataIos;
-    // **BUT**
-    // The above RawKeyEventData did not seem to work on Web when using an
-    // iPad+Safari and an Apple 10.5 Pro iPad keyboard, isRawKeyIos did not
-    // become true!
-    // So CMD modifier did not get used, only CTRL worked. We can use context
-    // based Theme.platform instead here and skip RawKeyEventData, or just
-    // combine it with RawKeyEventData. Since we have a context it works too.
-    final TargetPlatform platform = Theme.of(context).platform;
-    // Should COMMAND modifier be used instead of CTRL for keyboard COPY-PASTE?
-    // Use all sources we have to determine if it is iOS or macOS that should
-    // use CMD for copy/paste instead of CTRL.
-    final bool useCommandModifier = isRawKeyMacOS ||
-        isRawKeyIos ||
-        platform == TargetPlatform.iOS ||
-        platform == TargetPlatform.macOS;
-
-    // isModifierPressed will be true when COMMAND key is pressed on macOS/iOS
-    // OR when CTRL key is pressed on other platforms.
-    final bool isModifierPressed =
-        useCommandModifier ? event.isMetaPressed : event.isControlPressed;
-
-    // The raw keyboard listener reacts to both up and down events, we only use
-    // down, so that we only execute the copy and paste keyboard command once
-    // when the keys are pressed down. We do not want to do it 2nd time when
-    // the key goes up.
-    if (event.runtimeType == RawKeyDownEvent) {
-      // If logical key is paste OR modifier+V and we use ctrlV paste behavior,
-      // we get the latest clipboard data.
-      if ((event.logicalKey == LogicalKeyboardKey.paste ||
-              (isModifierPressed &&
-                  event.logicalKey == LogicalKeyboardKey.keyV)) &&
-          widget.copyPasteBehavior.ctrlV) {
-        unawaited(_getClipboard());
-      }
-      // If logical key is copy or modifier+C and we used ctrlC copy behavior,
-      // we set the current color to the clipboard data.
-      if ((event.logicalKey == LogicalKeyboardKey.copy ||
-              (isModifierPressed &&
-                  event.logicalKey == LogicalKeyboardKey.keyC)) &&
-          widget.copyPasteBehavior.ctrlC) {
-        unawaited(_setClipboard());
-      }
-    }
   }
 
   // Get the current clipboard data. Try to parse it to a Color object.
@@ -2208,17 +2357,18 @@ class _ColorPickerState extends State<ColorPicker> {
       // pasting when the edit field is not active.
       await Future<void>.delayed(const Duration(milliseconds: 100));
       setState(() {
-        // If opacity is not enabled, we remove any alpha from pasted colors.
-        _selectedColor =
-            widget.enableOpacity ? clipColor : clipColor.withAlpha(0xFF);
-        _opacity = widget.enableOpacity ? _selectedColor.opacity : 1;
+        // We always remove any alpha from pasted colors.
+        _selectedColor = clipColor.withAlpha(0xFF);
+        // If opacity is enabled, we capture the opacity from the pasted color.
+        _opacity = widget.enableOpacity ? clipColor.opacity : 1;
+        _addToRecentColors(_selectedColor.withOpacity(_opacity));
         // Color changed outside wheel and edit field, a new shade or color was
         // selected outside the wheel and edit, they should update!
         _wheelShouldUpdate = true;
         _editShouldUpdate = true;
         // Make a swatch of the new via paste _selectedColor for the wheel.
         _typeToSwatchMap[ColorPickerType.wheel] = <ColorSwatch<Object>>[
-          ColorTools.createPrimarySwatch(_selectedColor.withAlpha(0xFF)),
+          ColorTools.createPrimarySwatch(_selectedColor),
         ];
         // Move the picker to the pasted color value and update active swatch.
         _findPicker();
@@ -2284,19 +2434,14 @@ class _ColorPickerState extends State<ColorPicker> {
     switch (widget.copyPasteBehavior.copyFormat) {
       case ColorPickerCopyFormat.dartCode:
         colorString = '0x${_selectedColor.hexAlpha}';
-        break;
       case ColorPickerCopyFormat.hexRRGGBB:
         colorString = _selectedColor.hex;
-        break;
       case ColorPickerCopyFormat.hexAARRGGBB:
         colorString = _selectedColor.hexAlpha;
-        break;
       case ColorPickerCopyFormat.numHexRRGGBB:
         colorString = '#${_selectedColor.hex}';
-        break;
       case ColorPickerCopyFormat.numHexAARRGGBB:
         colorString = '#${_selectedColor.hexAlpha}';
-        break;
     }
     final ClipboardData data = ClipboardData(text: colorString);
     await Clipboard.setData(data);
