@@ -11,7 +11,7 @@ Future<bool> colorPickerDialog(
   WidgetRef ref, {
   bool cardRemote = false,
 }) async {
-  final ColorScheme colorScheme = Theme.of(context).colorScheme;
+  final ThemeData theme = Theme.of(context);
   return ColorPicker(
     color: cardRemote
         ? ref.watch(cardPickerColorPod)
@@ -67,10 +67,11 @@ Future<bool> colorPickerDialog(
       closeButton: ref.watch(closeButtonPod),
       closeIsLast: ref.watch(closeIsLastPod),
       dialogActionButtons: ref.watch(dialogActionButtonsPod),
+      dialogActionOnlyOkButton: ref.watch(dialogActionOnlyOkButtonPod),
       dialogActionOrder: ref.watch(dialogActionsOrderPod),
       dialogActionIcons: ref.watch(dialogActionIconsPod),
-      dialogOkButtonType: ColorPickerActionButtonType.outlined,
-      dialogCancelButtonType: ColorPickerActionButtonType.text,
+      dialogOkButtonType: ColorPickerActionButtonType.filled,
+      dialogCancelButtonType: ColorPickerActionButtonType.filledTonal,
     ),
     width: ref.watch(sizePod),
     height: ref.watch(sizePod),
@@ -81,6 +82,7 @@ Future<bool> colorPickerDialog(
     hasBorder: ref.watch(hasBorderPod),
     borderRadius: ref.watch(borderRadiusPod),
     columnSpacing: ref.watch(columnSpacingPod),
+    toolbarSpacing: 0,
     wheelDiameter: ref.watch(wheelDiameterPod),
     wheelWidth: ref.watch(wheelWidthPod),
     wheelHasBorder: ref.watch(wheelHasBorderPod),
@@ -88,47 +90,51 @@ Future<bool> colorPickerDialog(
     wheelSquareBorderRadius: ref.watch(wheelSquareBorderRadiusPod),
     enableTooltips: ref.watch(enableTooltipsPod),
     pickersEnabled: ref.watch(pickersEnabledPod),
-    selectedPickerTypeColor: colorScheme.primary,
+    pickerTypeLabels: const <ColorPickerType, String>{
+      ColorPickerType.both: 'P & A',
+      ColorPickerType.bw: 'B & W',
+    },
+    selectedPickerTypeColor: theme.colorScheme.primary,
     title: ref.watch(showTitlePod)
         ? Text(
             'ColorPicker',
-            style: Theme.of(context).textTheme.titleLarge,
+            style: theme.textTheme.titleLarge,
           )
         : null,
     heading: ref.watch(showHeadingPod)
         ? Text(
             'Select color',
-            style: Theme.of(context).textTheme.headlineSmall,
+            style: theme.textTheme.headlineSmall,
           )
         : null,
     subheading: ref.watch(showSubheadingPod)
         ? Text(
             'Select color shade',
-            style: Theme.of(context).textTheme.titleMedium,
+            style: theme.textTheme.titleMedium,
           )
         : null,
     tonalSubheading: ref.watch(showTonalSubheadingPod)
         ? Text(
             'Material 3 tonal palette',
-            style: Theme.of(context).textTheme.titleMedium,
+            style: theme.textTheme.titleMedium,
           )
         : null,
     wheelSubheading: ref.watch(showSubheadingPod)
         ? Text(
             'Selected color and its color swatch',
-            style: Theme.of(context).textTheme.titleMedium,
+            style: theme.textTheme.titleMedium,
           )
         : null,
     opacitySubheading: ref.watch(showOpacitySubheadingPod)
         ? Text(
             'Opacity',
-            style: Theme.of(context).textTheme.titleMedium,
+            style: theme.textTheme.titleMedium,
           )
         : null,
     recentColorsSubheading: ref.watch(showRecentSubheadingPod)
         ? Text(
             'Recent colors',
-            style: Theme.of(context).textTheme.titleMedium,
+            style: theme.textTheme.titleMedium,
           )
         : null,
     showMaterialName: ref.watch(showMaterialNamePod),
@@ -143,6 +149,7 @@ Future<bool> colorPickerDialog(
         : ref.watch(dialogRecentColorsPod),
     maxRecentColors: cardRemote ? 8 : 5,
     customColorSwatchesAndNames: App.colorsNameMap,
+    customSecondaryColorSwatchesAndNames: App.colorsOptionsMap,
   ).showPickerDialog(
     context,
     elevation: 2,
