@@ -9,15 +9,15 @@ class App {
   // Web demo with inside the app. Shown on the start screen in the demo,
   // so people testing it don't have to ask. Also info for the About screen.
   static const String appName = 'FlexColorPicker';
-  static const String version = '3.3.1';
+  static const String version = '3.4.0';
   static const String packageVersion = 'FlexColorPicker package $version';
   static final Uri packageUri = Uri(
     scheme: 'https',
     host: 'pub.dev',
     path: 'packages/flex_color_picker',
   );
-  static const String flutterVersion = 'Channel stable 3.16.8 (canvaskit)';
-  static const String copyright = '© 2020 - 2023';
+  static const String flutterVersion = 'Channel stable 3.19.2 (canvaskit)';
+  static const String copyright = '© 2020 - 2024';
   static const String author = 'Mike Rydstrom';
   static const String license = 'BSD 3-Clause License';
   static const String icon = 'assets/images/app_icon.png';
@@ -55,6 +55,29 @@ class App {
   static const Color blueJean = Color(0xFF4f75b8);
   static const Color deepBlueSea = Color(0xFF132b80);
 
+  static const Color mojo = Color(0xFFC74141);
+  static const Color trendyPink = Color(0xFF9565A5);
+  static const Color parsley = Color(0xFF16552C);
+  static const Color walnut = Color(0xFF753D1F);
+  static const Color parsleyOpacity = Color(0x5516552C);
+
+  // Add a custom white to black grey scale.
+  static const MaterialColor whiteBlueBlack = MaterialColor(
+    0xFF4355B9, // Set the 500 index value here.
+    <int, Color>{
+      50: Color(0xFFFFFFFF),
+      100: Color(0xFFF0EFFF),
+      200: Color(0xFFBAC3FF),
+      300: Color(0xFF7789F0),
+      400: Color(0xFF5D6FD4),
+      500: Color(0xFF4355B9),
+      600: Color(0xFF293CA0),
+      700: Color(0xFF08218A),
+      800: Color(0xFF00105C),
+      900: Color(0xFF000000),
+    },
+  );
+
   // Add a custom white to black grey scale.
   static const MaterialColor whiteToBlack = MaterialColor(
     0xFF7C7D80, // Set the 500 index value here.
@@ -72,20 +95,41 @@ class App {
     },
   );
 
-  // Add a custom white to black grey scale.
-  static const MaterialColor whiteBlueBlack = MaterialColor(
-    0xFF4355B9, // Set the 500 index value here.
+  // Add a custom black transparency.
+  //
+  // Index 50 is fully transparent black.
+  //
+  // Index 100 is 10% of max alpha value or 10% opacity, and so on.
+  static const MaterialColor blackTransparency = MaterialColor(
+    0x7F000000, // Set the 500 index value here.
     <int, Color>{
-      50: Color(0xFFFFFFFF),
-      100: Color(0xFFF0EFFF),
-      200: Color(0xFFBAC3FF),
-      300: Color(0xFF7789F0),
-      400: Color(0xFF5D6FD4),
-      500: Color(0xFF4355B9),
-      600: Color(0xFF293CA0),
-      700: Color(0xFF08218A),
-      800: Color(0xFF00105C),
-      900: Color(0xFF000000),
+      50: Color(0x00000000),
+      100: Color(0x19000000),
+      200: Color(0x33000000),
+      300: Color(0x4C000000),
+      400: Color(0x66000000),
+      500: Color(0x7F000000),
+      600: Color(0x99000000),
+      700: Color(0xB2000000),
+      800: Color(0xCC000000),
+      900: Color(0xE5000000),
+    },
+  );
+
+  // Add a custom parsley swatch
+  static const MaterialColor allSwatchParsleyTransparent = MaterialColor(
+    0x9930874C, // Set the 500 index value here.
+    <int, Color>{
+      50: Color(0x99F5FFF2),
+      100: Color(0x99ADF3B9),
+      200: Color(0x9982D995),
+      300: Color(0x9967BD7C),
+      400: Color(0x994CA164),
+      500: Color(0x9930874C),
+      600: Color(0x990C6D35),
+      700: Color(0x99005226),
+      800: Color(0x99003918),
+      900: Color(0x9900210B),
     },
   );
 
@@ -113,5 +157,16 @@ class App {
         ColorTools.createPrimarySwatch(deepBlueSea): 'Deep blue sea',
         whiteBlueBlack: 'White via Blue to Black',
         whiteToBlack: 'White to black',
+        blackTransparency: 'Black transparency',
+      };
+
+  static Map<ColorSwatch<Object>, String> get colorsOptionsMap =>
+      <ColorSwatch<Object>, String>{
+        ColorTools.createPrimarySwatch(mojo): 'Mojo',
+        ColorTools.createPrimarySwatch(trendyPink): 'Trendy pink',
+        ColorTools.createPrimarySwatch(parsley): 'Parsley',
+        ColorTools.createPrimarySwatch(walnut): 'Walnut',
+        ColorTools.createPrimarySwatch(parsleyOpacity): 'Transparent parsley',
+        allSwatchParsleyTransparent: 'Parsley all indexes transparent',
       };
 }
