@@ -14,7 +14,6 @@ class TonalPaletteColors extends StatefulWidget {
     Key? key,
     required this.spacing,
     required this.runSpacing,
-    required this.columnSpacing,
     required this.selectedColor,
     required this.onSelectColor,
     required this.tonalShouldUpdate,
@@ -33,9 +32,6 @@ class TonalPaletteColors extends StatefulWidget {
 
   /// The run spacing between the color pick items when wrapped on several rows.
   final double runSpacing;
-
-  /// The spacing after the main colors.
-  final double columnSpacing;
 
   /// The selected color.
   final Color selectedColor;
@@ -102,31 +98,28 @@ class _TonalPaletteColorsState extends State<TonalPaletteColors> {
   Widget build(BuildContext context) {
     final double effectiveBorderRadius =
         widget.borderRadius ?? widget.width / 4.0;
-    return Padding(
-      padding: EdgeInsets.only(bottom: widget.columnSpacing),
-      child: Wrap(
-        spacing: widget.spacing,
-        runSpacing: widget.runSpacing,
-        children: <Widget>[
-          for (final Color color in tonalColors)
-            ColorIndicator(
-              isSelected: widget.selectedColor == color ||
-                  widget.selectedColor.value == color.value,
-              color: color,
-              width: widget.width,
-              height: widget.height,
-              borderRadius: effectiveBorderRadius,
-              hasBorder: widget.hasBorder,
-              borderColor: widget.borderColor,
-              elevation: widget.elevation,
-              selectedIcon: widget.selectedColorIcon,
-              onSelect: () {
-                widget.onSelectColor(color);
-              },
-              selectedRequestsFocus: widget.selectedRequestsFocus,
-            ),
-        ],
-      ),
+    return Wrap(
+      spacing: widget.spacing,
+      runSpacing: widget.runSpacing,
+      children: <Widget>[
+        for (final Color color in tonalColors)
+          ColorIndicator(
+            isSelected: widget.selectedColor == color ||
+                widget.selectedColor.value == color.value,
+            color: color,
+            width: widget.width,
+            height: widget.height,
+            borderRadius: effectiveBorderRadius,
+            hasBorder: widget.hasBorder,
+            borderColor: widget.borderColor,
+            elevation: widget.elevation,
+            selectedIcon: widget.selectedColorIcon,
+            onSelect: () {
+              widget.onSelectColor(color);
+            },
+            selectedRequestsFocus: widget.selectedRequestsFocus,
+          ),
+      ],
     );
   }
 }
